requirements.txt
flask==2.3.2 
mysql-connector-python==8.0.33 
werkzeug==2.3.6 
bootstrap-flask==2.3.3 
python-dotenv==1.0.0 
